Move saga tests to separate file
